Refactor: 리뷰반영 MockUp 파일 생성(#4)
