+++ src/Utils/index.tsx
@@ -1,2 +1,3 @@
 export { LOCAL_STORAGE } from "./Constants/constants";
 export { default as STATUS } from "./Constants/TodoStatus";
+export { default as MockUp } from "./Constants/MockUp";
